Moved config stuff to a new config package
