+++ src/main/java/fr/smyler/mcct/config/ConfigFloatValue.java
@@ -1,7 +1,11 @@
-package fr.smyler.mcct.tweaks.config;
+package fr.smyler.mcct.config;
+
+import java.lang.reflect.Type;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
 
 public class ConfigFloatValue extends ConfigNumericValue<Float> {
 
@@ -12,6 +16,15 @@ public class ConfigFloatValue extends ConfigNumericValue<Float> {
 	@Override
 	public JsonElement toJsonPrimitive() {
 		return new JsonPrimitive(this.get());
+	}
+
+}
+
+class ConfigFloatValueSerializer implements JsonSerializer<ConfigFloatValue>{
+
+	@Override
+	public JsonElement serialize(ConfigFloatValue src, Type typeOfSrc, JsonSerializationContext context) {
+		return src.toJsonPrimitive();
 	}
 
 }
