+++ src/main/java/fr/smyler/mcct/config/ConfigIntValue.java
@@ -1,7 +1,11 @@
-package fr.smyler.mcct.tweaks.config;
+package fr.smyler.mcct.config;
+
+import java.lang.reflect.Type;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
 
 public class ConfigIntValue extends ConfigNumericValue<Integer> {
 
@@ -12,6 +16,15 @@ public class ConfigIntValue extends ConfigNumericValue<Integer> {
 	@Override
 	public JsonElement toJsonPrimitive() {
 		return new JsonPrimitive(this.get());
+	}
+
+}
+
+class ConfigIntValueSerializer implements JsonSerializer<ConfigIntValue>{
+
+	@Override
+	public JsonElement serialize(ConfigIntValue src, Type typeOfSrc, JsonSerializationContext context) {
+		return src.toJsonPrimitive();
 	}
 
 }
